fix: preserve PPTX borders during import
